requirements.txt
Django==5.2.4
djangorestframework==3.15.2
djoser==2.2.3
django-cors-headers==4.3.1
django-import-export==4.1.1
djangorestframework-simplejwt==5.3.0
Pillow==10.4.0
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
psycopg2-binary==2.9.9
